refactored typo in screens
